% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance matrix between BBS and KOH}
\usage{
distance()
}
\value{
A matrix.
}
\description{
Create a matrix containing distances from every site of BBS data to every site of KOH data
}
\examples{
distance ()
}
