% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_data.R
\name{get_all_data}
\alias{get_all_data}
\title{Get all the data from Goole account.}
\usage{
get_all_data()
}
\value{
6 variables
\itemize{
\item {\code{d0914}}
{Average densities of each bird between 2009 and 2014 from BBS data}
\item \code{d1520}
{Average densities of each bird between 2015 and 2020 from BBS data}
\item \code{dkoh}
{Average densities of each bird from KOH data}
\item \code{site_bbs}
{Locations and elevations of sites from BBS data}
\item \code{site_bbs0914}
{Locations and elevations of sites from BBS data between 2009 and 2014}
\item \code{site_bbs1520}
{Locations and elevations of sites from BBS data between 2015 and 2020}
\item \code{site_koh}
{Locations and elevations of sites from KOH data}
}
}
\description{
Get all the data from Goole account.
}
\examples{
get_all_data ()
}
