% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance.R
\name{get_distance}
\alias{get_distance}
\title{Distance matrix between BBS and KOH}
\usage{
get_distance()
}
\value{
A matrix.
}
\description{
Create a matrix containing distances from every site of BBS data to every site of KOH data
}
\examples{
get_distance ()
}
